requirements.txt
google-generativeai
python-dotenv
google-ai-generativelanguage
requests
beautifulsoup4
duckduckgo-search
charset_normalizer